02-use-cases/beginner/multi_agents/requirements.txt
veadk-python==0.2.26
agentkit-sdk-python==0.1.13